# hot-wheels-backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Hot Wheels Collector API
  version: 1.0.0
  description: |
    # Hot Wheels Collector Platform API

    A comprehensive REST API for Hot Wheels collectors to manage their collections,
    trade with other collectors, and explore the complete Hot Wheels catalog.

    ## Features
    - **Authentication**: JWT-based secure authentication system
    - **Catalog**: Browse 50+ years of Hot Wheels models
    - **Collections**: Track and manage personal collections
    - **Marketplace**: Buy, sell, and trade with other collectors
    - **Reviews**: Rate and review transactions

    ## Getting Started
    1. Register a new account via `/api/v1/auth/register`
    2. Login to receive JWT tokens via `/api/v1/auth/login`
    3. Include the access token in the `Authorization` header
    4. Start exploring the catalog and managing your collection!

    ## Rate Limits
    - Anonymous users: 100 requests/minute
    - Authenticated users: 500 requests/minute
    - Rate limit headers included in all responses

    ## Support
    For API support, please contact the development team.

  contact:
    name: API Support
    email: support@hotwheels-collector.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.hotwheels-collector.com
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token obtained from `/api/v1/auth/login`.
        Include in Authorization header: `Bearer <token>`
  schemas:
...
    Error:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            message:
              type: string
              example: "Invalid request parameters"
            code:
              type: string
              example: "VALIDATION_ERROR"
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total:
          type: integer
          example: 150
        total_pages:
          type: integer
          example: 8
        has_next:
          type: boolean
          example: true
        has_prev:
          type: boolean
          example: false

    # ==================== User Schemas ====================
...
    CarModel:
      type: object
      properties:
        model_id:
          type: integer
          example: 1
        casting_id:
          type: integer
          example: 1
        series_id:
          type: integer
          nullable: true
          example: 1
        release_year:
          type: integer
          example: 2024
        color:
          type: string
          example: "Spectraflame Red"
        tampo_design:
          type: string
          nullable: true
          example: "Racing stripes with #7"
        wheel_type:
          type: string
          nullable: true
          example: "5-Spoke"
        base_color:
          type: string
          nullable: true
          example: "Chrome"
        window_color:
          type: string
          nullable: true
          example: "Tinted"
        interior_color:
          type: string
          nullable: true
          example: "Black"
        production_code:
          type: string
          nullable: true
          example: "L2593"
        sku:
          type: string
          nullable: true
          example: "HCW35-N9B1"
        rarity_level:
          type: string
          enum: [Common, Uncommon, Rare, Chase, Super Treasure Hunt]
          example: "Common"
        estimated_production_quantity:
          type: integer
          nullable: true
          example: 50000
        msrp:
          type: number
          format: float
          nullable: true
          example: 1.29
        casting_name:
          type: string
          description: Included when relations are requested
          example: "'67 Camaro"
        series_name:
          type: string
          nullable: true
          description: Included when relations are requested
          example: "Mainline 2024"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Listing:
      type: object
      properties:
        listing_id:
          type: integer
          example: 1
        seller_id:
          type: integer
          example: 1
        collection_item_id:
          type: integer
          nullable: true
          example: 5
        model_id:
          type: integer
          example: 1
        listing_type:
          type: string
          enum: [sale, trade, auction]
          example: "sale"
        price:
          type: number
          format: float
          nullable: true
          example: 15.99
          description: Required for sale listings, nullable for trades
        condition:
          type: string
          enum: [Mint, Near Mint, Excellent, Good, Fair, Poor]
          example: "Mint"
        description:
          type: string
          nullable: true
          example: "Pristine condition, never opened"
        status:
          type: string
          enum: [active, sold, cancelled, expired]
          example: "active"
        views_count:
          type: integer
          example: 42
        seller:
          type: object
          properties:
            username:
              type: string
              example: "hotwheels_fan"
            is_verified:
              type: boolean
              example: false
          description: Included when relations are requested
        car_model:
          $ref: '#/components/schemas/CarModel'
          description: Included when relations are requested
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
          nullable: true
        sold_at:
          type: string
          format: date-time
          nullable: true
    CreateListingRequest:
      type: object
      required:
        - model_id
        - listing_type
        - condition
      properties:
        model_id:
          type: integer
          example: 1
        listing_type:
          type: string
          enum: [sale, trade, auction]
          example: "sale"
        condition:
          type: string
          enum: [Mint, Near Mint, Excellent, Good, Fair, Poor]
          example: "Mint"
        price:
          type: number
          format: float
          example: 15.99
          description: Required for sale listings
        description:
          type: string
          example: "Pristine condition, never opened"
        collection_item_id:
          type: integer
          example: 5
          description: Link to your collection item
        expires_at:
          type: string
          format: date-time
          example: "2024-12-31T23:59:59Z"
    UpdateListingRequest:
      type: object
      properties:
        price:
          type: number
          format: float
          example: 14.99
        description:
          type: string
          example: "Updated description"
        condition:
          type: string
          enum: [Mint, Near Mint, Excellent, Good, Fair, Poor]
    Transaction:
      type: object
      properties:
        transaction_id:
          type: integer
          example: 1
        listing_id:
          type: integer
          example: 1
        buyer_id:
          type: integer
          example: 2
        seller_id:
          type: integer
          example: 1
        transaction_type:
          type: string
          enum: [purchase, trade]
          example: "purchase"
        amount:
          type: number
          format: float
          example: 15.99
        payment_method:
          type: string
          nullable: true
          example: "PayPal"
        payment_status:
          type: string
          enum: [pending, completed, failed, refunded]
          example: "pending"
        shipping_status:
          type: string
          enum: [not_shipped, shipped, in_transit, delivered]
          example: "not_shipped"
        tracking_number:
          type: string
          nullable: true
          example: "1Z999AA10123456784"
        buyer_username:
          type: string
          description: Included when relations are requested
          example: "buyer123"
        seller_username:
          type: string
          description: Included when relations are requested
          example: "seller456"
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
    CreateTransactionRequest:
      type: object
      required:
        - listing_id
      properties:
        listing_id:
          type: integer
          example: 1
        payment_method:
          type: string
          example: "PayPal"
    Review:
      type: object
      properties:
        review_id:
          type: integer
          example: 1
        transaction_id:
          type: integer
          example: 1
        reviewer_id:
          type: integer
          example: 2
        reviewee_id:
          type: integer
          example: 1
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
        comment:
          type: string
          nullable: true
          example: "Great seller, fast shipping!"
        reviewer_username:
          type: string
          description: Included when relations are requested
          example: "buyer123"
        reviewee_username:
          type: string
          description: Included when relations are requested
          example: "seller456"
        created_at:
          type: string
          format: date-time
    CreateReviewRequest:
      type: object
      required:
        - transaction_id
        - reviewee_id
        - rating
      properties:
        transaction_id:
          type: integer
          example: 1
        reviewee_id:
          type: integer
          example: 1
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
        comment:
          type: string
          example: "Great seller, fast shipping!"
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              message: "Missing required fields: email, password"
    Unauthorized:
      description: Unauthorized - authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              message: "Authentication required"
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              message: "Access forbidden"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              message: "Resource not found"
    Conflict:
      description: Conflict - resource already exists or invalid state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              message: "Email already registered"
...
paths:
...
  /api/v1/marketplace/listings:
    get:
      summary: Get marketplace listings
      description: |
        Returns active marketplace listings with optional filters.

        **Filtering Options:**
        - By car model ID
        - By listing type (sale/trade/auction)
        - By seller
        - By price range
      tags:
        - Marketplace
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
        - name: model_id
          in: query
          description: Filter by car model
          schema:
            type: integer
        - name: listing_type
          in: query
          description: Filter by listing type
          schema:
            type: string
            enum: [sale, trade, auction]
        - name: seller_id
          in: query
          description: Filter by seller
          schema:
            type: integer
        - name: price_min
          in: query
          description: Minimum price
          schema:
            type: number
            format: float
        - name: price_max
          in: query
          description: Maximum price
          schema:
            type: number
            format: float
      responses:
        '200':
          description: Listings
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      listings:
                        type: array
                        items:
                          $ref: '#/components/schemas/Listing'
                      pagination:
                        $ref: '#/components/schemas/PaginationMeta'
    post:
      summary: Create a listing
      description: Creates a new marketplace listing
      tags:
        - Marketplace
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateListingRequest'
            examples:
              sale:
                summary: Sale listing
                value:
                  model_id: 1
                  listing_type: "sale"
                  condition: "Mint"
                  price: 15.99
                  description: "Pristine condition, never opened"
              trade:
                summary: Trade listing
                value:
                  model_id: 1
                  listing_type: "trade"
                  condition: "Near Mint"
                  description: "Looking to trade for Chase variants"
      responses:
        '201':
          description: Listing created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      listing:
                        $ref: '#/components/schemas/Listing'
                  message:
                    type: string
                    example: "Listing created successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/marketplace/listings/{listing_id}:
    get:
      summary: Get a listing
      description: Returns detailed information about a specific listing
      tags:
        - Marketplace
      parameters:
        - name: listing_id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Listing details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      listing:
                        $ref: '#/components/schemas/Listing'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update a listing
      description: Updates an existing listing (seller only)
      tags:
        - Marketplace
      security:
        - BearerAuth: []
      parameters:
        - name: listing_id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateListingRequest'
      responses:
        '200':
          description: Listing updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      listing:
                        $ref: '#/components/schemas/Listing'
                  message:
                    type: string
                    example: "Listing updated successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/marketplace/listings/{listing_id}/cancel:
    post:
      summary: Cancel a listing
      description: Cancels an active listing (seller only)
      tags:
        - Marketplace
      security:
        - BearerAuth: []
      parameters:
        - name: listing_id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Listing cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Listing cancelled successfully"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/marketplace/transactions:
    post:
      summary: Create a transaction
      description: |
        Initiates a purchase or trade transaction.

        **Important:** You cannot purchase from yourself.
      tags:
        - Marketplace
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransactionRequest'
      responses:
        '201':
          description: Transaction created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      transaction:
                        $ref: '#/components/schemas/Transaction'
                  message:
                    type: string
                    example: "Transaction completed successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/marketplace/transactions/purchases:
    get:
      summary: Get user's purchases
      description: Returns the authenticated user's purchase history
      tags:
        - Marketplace
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Purchase history
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      transactions:
                        type: array
                        items:
                          $ref: '#/components/schemas/Transaction'
                      pagination:
                        $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/marketplace/transactions/sales:
    get:
      summary: Get user's sales
      description: Returns the authenticated user's sales history
      tags:
        - Marketplace
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Sales history
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      transactions:
                        type: array
                        items:
                          $ref: '#/components/schemas/Transaction'
                      pagination:
                        $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/marketplace/reviews:
    post:
      summary: Create a review
      description: |
        Creates a review for a completed transaction.

        **Rules:**
        - Rating must be between 1 and 5
        - Can only review once per transaction
        - Must be a participant in the transaction
      tags:
        - Marketplace
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewRequest'
      responses:
        '201':
          description: Review created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      review:
                        $ref: '#/components/schemas/Review'
                  message:
                    type: string
                    example: "Review submitted successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'